Generate ComInterfaceExternalProxyStubElements only in assembly manifests
This is because this information is needed by client software other than
VB6 (which always uses IDispatch-based proxying). So VB6 programs will
work without this information at all (so no point in putting in it the
program manifests for VB6 exe's.

However there may still be a problem where the same interface is used in
multiple libraries referenced from different manifests, which I think
Windows doesn't like, but this needs further investigation,. So this may
no be the last change in this area!